+++ GRDB/Core/Support/Foundation/Data.swift
@@ -8,8 +8,8 @@ import Foundation
 /// Data is convertible to and from DatabaseValue.
 extension Data : DatabaseValueConvertible, StatementColumnConvertible {
     public init(sqliteStatement: SQLiteStatement, index: Int32) {
-        if let bytes = sqlite3_column_blob(sqliteStatement, Int32(index)) {
-            let count = Int(sqlite3_column_bytes(sqliteStatement, Int32(index)))
+        if let bytes = sqlite3_column_blob(sqliteStatement, index) {
+            let count = Int(sqlite3_column_bytes(sqliteStatement, index))
             self.init(bytes: bytes, count: count) // copy bytes
         } else {
             self.init()
